add 배치 수정, 색상 적용
전체 배치도 수정
요소별 색상 수정
전체 타이머 제거
